setup with tailwind, hero section and navbar done
