+++ open-giving/package.json
@@ -23,13 +23,18 @@
     "non-profit-donations": "^0.1.2",
     "qrcode.react": "^4.0.1",
     "react": "^18",
+    "react-device-detect": "^2.2.3",
     "react-dom": "^18",
+    "react-icons": "^5.3.0",
     "stylis": "^4.3.0",
     "sweetalert2": "^11.9.0",
     "use-local-storage-state": "^19.1.0"
   },
   "devDependencies": {
+    "autoprefixer": "^10.4.20",
     "eslint": "^8",
-    "eslint-config-next": "14.0.1"
+    "eslint-config-next": "14.0.1",
+    "postcss": "^8.4.47",
+    "tailwindcss": "^3.4.14"
   }
 }
