+++ open-giving/src/components/Header.js
@@ -1,21 +1,26 @@
-import { Typography } from '@mui/material';
+import Image from 'next/image';
+import Link from 'next/link';
 import React from 'react';
+import { isMobile } from 'react-device-detect';
+import { FaGithub } from 'react-icons/fa';
 
 const Header = () => (
-  <>
-    <Typography variant="h3" component="h1" align="center" color="text.primary">
-      Open Giving
-    </Typography>
-    <Typography
-      variant="subtitle1"
-      component="h2"
-      gutterBottom
-      align="center"
-      color="text.primary"
+  <div className="py-8 px-6 md:px-12 2xl:px-[20%] h-20 w-full border-b border-[#AE6DEA45] flex flex-row items-center justify-between">
+    <Image
+      src={'/logo.svg'}
+      alt="Open Giving Logo"
+      width={isMobile ? 180 : 220}
+      height={isMobile ? 35 : 50}
+    />
+    <Link
+      href="https://github.com/Chimoney/non-profit-donations"
+      target="_blank"
+      rel="noopener noreferrer"
+      aria-label="github repository"
     >
-      Donate to Verified Non-Profits
-    </Typography>
-  </>
+      <FaGithub className="size-[25px] md:size-[35px]" />
+    </Link>
+  </div>
 );
 
 export default Header;
