+++ open-giving/src/components/constants/index.js
@@ -1,0 +1,48 @@
+export const HeroContent = {
+  header: 'Donate to Verified Non-Profits',
+  subheader:
+    'Support not-for-profit initiatives who’s activities will lead to a more prosperous world',
+};
+
+export const donationMethods = [
+  {
+    method:'Chimoney',
+    icon: '/new/donation/chimoney.svg',
+  },
+  {
+    method:'Interledger',
+    icon: '/new/donation/interledger.svg',
+  },
+  {
+    method:'Paypal',
+    icon: '/new/donation/paypal.svg',
+  },
+  {
+    method:'Stripe',
+    icon: '/new/donation/stripe.svg',
+  },
+  {
+    method:'Venmo',
+    icon: '/new/donation/venmo.svg',
+  },
+  {
+    method:'Cashapp',
+    icon: '/new/donation/cashapp.svg',
+  },
+  {
+    method:'Patreon',
+    icon: '/new/donation/patreon.svg',
+  },
+  {
+    method:'Donorbox',
+    icon: '/new/donation/donorbox.svg',
+  },
+  {
+    method:'Donation Link',
+    icon: '/new/donation/donationlink.svg',
+  },
+  {
+    method:'Stablecoin',
+    icon: '/new/donation/stablecoin.svg',
+  },
+]
